vida de int a double

+++ src/magosVsMortifagos/Estudiante.java
@@ -1,7 +1,7 @@
 package magosVsMortifagos;
 
 public class Estudiante extends Mago {
-	public Estudiante(String nombre, int nivelMagia, int puntosVida) {
+	public Estudiante(String nombre, int nivelMagia, double puntosVida) {
 		super(nombre, nivelMagia, puntosVida, Varita.PRINCIPIANTE);
 	}
 }
