+++ src/magosVsMortifagos/Mago.java
@@ -1,7 +1,7 @@
 package magosVsMortifagos;
 
 public class Mago extends Personaje{
-	public Mago(String nombre, int nivelMagia, int puntosVida, Varita varita) {
+	public Mago(String nombre, int nivelMagia, double puntosVida, Varita varita) {
 		super(nombre, nivelMagia, puntosVida, varita);
 	}
