+++ src/magosVsMortifagos/Profesor.java
@@ -1,7 +1,7 @@
 package magosVsMortifagos;
 
 public class Profesor extends Mago{
-	public Profesor(String nombre, int nivelMagia, int puntosVida) {
+	public Profesor(String nombre, int nivelMagia, double puntosVida) {
 		super(nombre, nivelMagia, puntosVida, Varita.PROFESIONAL);
 	}
 }
